+++ PriceCalculatorKata/Product.cs
@@ -11,7 +11,7 @@ namespace PriceCalculatorKata
         private readonly IPriceCalculator _priceCalculator;
         private readonly IReporter _reporter;
         public Currency Currency { get; set; }
-        public Product(int UPC,float basePrice,IPriceCalculator priceCalculator,IReporter reporter,Currency? currency=null)
+        public Product(int UPC,double basePrice,IPriceCalculator priceCalculator,IReporter reporter,Currency? currency=null)
         {
             this.UPC = UPC;
             this.BasePrice = basePrice;
@@ -20,20 +20,20 @@ namespace PriceCalculatorKata
             Currency = currency??new Currency(){CurrencyCode = "USD"};
         }
         public int UPC { get; private set; }
-        private float _basePrice;
-        public float BasePrice
+        private double _basePrice;
+        public double BasePrice
         {
             get => _basePrice;
             set
             {
                 if (value > 0)
                 {
-                    _basePrice = (float)Math.Round(value, 2);
+                    _basePrice = (double)Math.Round(value, 2);
                 }
             }
         }
 
-        public float FinalPrice => _priceCalculator.Calculate(this);
+        public double FinalPrice => _priceCalculator.Calculate(this);
 
         public void Report(Action<string> displayMethod)
         {
